requirements.txt
gitpython
html2text
markdown
numpy
onnxruntime
requests
uvicorn
fastapi
flask
pydantic
prefect
transformers